multinotes-backend-llm-model-V2.0/commonai-backend-llm-model-V2.0/requirements.txt
adrf==0.1.6
aiohttp==3.9.5
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
argcomplete==1.10.3
asgiref==3.7.2
async-property==0.2.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.8.2
billiard==4.2.0
boto3==1.34.80
botocore==1.34.80
Brotli==1.1.0
CacheControl==0.14.0
cachetools==5.3.3
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
compressed_rtf==1.0.6
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.2
decorator==4.4.2
distro==1.9.0
Django==5.0.2
django-cors-headers==4.3.1
django-crontab==0.7.1
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
docx2txt==0.8
ebcdic==1.1.1
et_xmlfile==2.0.0
eval-type-backport==0.1.3
exceptiongroup==1.2.0
extract-msg==0.28.7
ffmpeg==1.4
ffprobe==0.5
filelock==3.13.1
firebase-admin==6.5.0
frozenlist==1.4.1
fsspec==2024.2.0
google-ai-generativelanguage==0.6.10
google-api-core==2.17.1
google-api-python-client==2.128.0
google-auth==2.28.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-generativeai==0.8.2
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.62.0
grpcio-status==1.62.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.21.2
hyperlink==21.0.0
idna==3.6
imageio==2.35.1
imageio-ffmpeg==0.5.1
IMAPClient==2.1.0
incremental==22.10.0
jmespath==1.0.1
kombu==5.3.6
lxml==5.3.0
markdown-it-py==3.0.0
mdurl==0.1.2
msgpack==1.0.8
multidict==6.0.5
mutagen==1.47.0
mysqlclient==2.2.4
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
openai==1.28.0
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.3
pdfminer.six==20191110
pillow==10.3.0
proglog==0.1.10
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
pyarrow==16.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.22
pycryptodome==3.21.0
pycryptodomex==3.21.0
pydantic==2.7.1
pydantic_core==2.18.2
pydub==0.25.1
pyfcm==1.5.4
Pygments==2.17.2
PyJWT==2.8.0
PyMuPDF==1.24.13
pyOpenSSL==24.1.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-pptx==0.6.23
pytubefix==8.1.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.3
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.1
rsa==4.9
s3transfer==0.10.1
service-identity==24.1.0
shellingham==1.5.4
six==1.12.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SpeechRecognition==3.8.1
sqlparse==0.4.4
sseclient==0.0.27
stripe==9.2.0
tabulate==0.9.0
textract==1.6.5
together==1.3.1
tqdm==4.66.2
trio==0.27.0
trio-websocket==0.11.1
Twisted==24.3.0
txaio==23.1.1
typer==0.12.3
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==13.1
wsproto==1.2.0
xlrd==1.2.0
XlsxWriter==3.2.0
yarl==1.9.4
yt-dlp==2024.10.7
zope.interface==6.3
